Add files via upload
Last version, used for the paper. Includes tests for the SF model in GEANT4.
Contains the notebook that was used to merge the pulses from the detectors, and its main output is the singles, doubles and triples rate.

Revert "Add files via upload"

This reverts commit a98eb70859f91efc4438345b3acff2b9b5c74260.

Add files via upload

Last version, used for the paper. Includes tests for the SF model in GEANT4.
Contains the notebook that was used to merge the pulses from the detectors, and its main output is the singles, doubles and triples rate.